Change the way server handles the clients.
The server will continuously receive messages from the client.
When the PTT is pressed the client will send a request to the server
for approval, if approved the client is able to speak, otherwise it will
be rejected. The request will be captured inside a JSON text to
distinguish it from the audio data.

Added some debugging print messages.

Fixed the issue of client only sent message after recording. Now it can
send message in some time interval.

Fixed the issue of multiple clients cannot connect.